refactor: extract starter visible driver family
